moving src to new folder structure, and updating so it still builds (shared at least)
