+++ Tests/SharedGoogleTests/environment_repository.cpp
@@ -12,7 +12,7 @@
 // 
 
 #include "pch.h"
-#include "../Shared/environment_repository_impl.h"
+#include <environment_repository_impl.h>
 
 using shared::infrastructure::environment_repository_impl;
