+++ Tests/SharedGoogleTests/process_service.cpp
@@ -12,7 +12,7 @@
 // 
 
 #include "pch.h"
-#include <../Shared/process_service_impl.h>
+#include <process_service_impl.h>
 #include <chrono>
 #include "Common.h"
