+++ src/shared/cpp.hint
